refactor: restrcuct directory
