--- src/data/roleplayFeedbackHistory.json
@@ -1,8 +1,0 @@
-[
-  { "id": "history-1", "date": "2025.11.15", "time": "10:30", "label": "Prod 핫픽스 대응" },
-  { "id": "history-2", "date": "2025.11.14", "time": "15:00", "label": "파트너 일정 조율" },
-  { "id": "history-3", "date": "2025.11.12", "time": "09:45", "label": "릴리즈 노트 공유" }
-]
-
-
-
